Backend/requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# AI & ML
openai==1.10.0
pinecone-client==3.0.2
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3

# OCR & Image Processing
pytesseract==0.3.10
Pillow==10.2.0
opencv-python==4.9.0.80



# Utilities
python-dateutil==2.8.2
httpx==0.26.0
